Finished input test data file option Fixed error in Algorithm.java , where LengthLabelList was not resetting resulting in incorrect output for multiple entries

+++ src/main/java/org/example/Test_Entry.java
@@ -4,8 +4,8 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class Test_Entry {
-    List<Float> points;
-    public Test_Entry(List<Float> points) {
+    List<Double> points;
+    public Test_Entry(List<Double> points) {
         this.points = points;
     }
 }
